Corrected the type in Task 1

+++ task-01/src/App.java
@@ -13,7 +13,7 @@ public class App {
 
         // 2
         System.out.println("Calculate expressions:");
-        int result1 = (46 + 10) * (10 / 3); 
+        double result1 = (46 + 10) * (10 / 3); 
         int result2 = 29 * 4 * -15;
         System.out.println(result1);
         System.out.println(result2);
@@ -21,7 +21,7 @@ public class App {
         // 3
         System.out.println("Calculate numbers:");
         int number = 10500;
-        int result3 = (number / 10) / 10;
+        double result3 = (number / 10) / 10;
         System.out.println(result3);
 
         // 4
